Refactor semester arg for test cases
