+++ src/test/java/seedu/address/testutil/TypicalModules.java
@@ -19,30 +19,30 @@ public class TypicalModules {
     public static final Module CS2040S = new ModuleBuilder()
             .withCode("CS2040S")
             .withYear("1")
-            .withSem("2")
+            .withSem("SEMESTER_2")
             .withGrade("A-").build();
     public static final Module CS2030S = new ModuleBuilder()
             .withCode("CS2030S")
             .withYear("1")
-            .withSem("2")
+            .withSem("SEMESTER_2")
             .withGrade("A").build();
     public static final Module MA2001 = new ModuleBuilder()
             .withCode("MA2001")
             .withYear("1")
-            .withSem("2")
+            .withSem("SEMESTER_2")
             .withGrade("B+").build();
 
     // Manually added
     public static final Module CS2101 = new ModuleBuilder()
             .withCode("MA2001")
             .withYear("2")
-            .withSem("1")
+            .withSem("SEMESTER_1")
             .withGrade("IP").build();
 
     public static final Module GEA1000 = new ModuleBuilder()
             .withCode("GEA1000")
             .withYear("1")
-            .withSem("1")
+            .withSem("SEMESTER_1")
             .withGrade("A+").build();
     //Add more
